able to load markers from hard coded coords
